requirements.txt
rdkit==2022.9.5
selfies==2.1.1
CairoSVG==2.7.0
matplotlib==3.7.1
pandas==2.0.1
numpy==1.24.3
seaborn==0.12.2
torch==2.0.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
botorch==0.8.4
evotorch==0.4.1
gpytorch==1.10